hdmi: throttle the render heartbeat log
"render frame=N style_loaded=1" every 300 frames is one line every five
seconds, 25k lines a day -- a third of this host's entire journal, to say
"still rendering". Every 18000 frames (~5 min) says the same thing.

This is not the frame-rate instrumentation: that is the [perf] block, which
MAPLIBRE_PERF already gates. Setting MAPLIBRE_PERF keeps the old fast
cadence, so the heartbeat is still there when watching a problem.

frame_count_ becomes uint64_t while here: at 60fps an int wraps in about
fourteen months, and signed overflow is undefined.

+++ hdmi/src/slint_map_gl.hpp
@@ -138,7 +138,9 @@ private:
     mbgl::Point<double> last_pos{};
     double min_zoom_ = 0.0;
     double max_zoom_ = 22.0;
-    int frame_count_ = 0;
+    // 64-bit: at 60fps an int wraps in about 14 months, and signed
+    // overflow is undefined -- this appliance is expected to just keep running.
+    uint64_t frame_count_ = 0;
     // flyTo duration; override with MAPLIBRE_FLY_MS. Deliberately long: V3D
     // renders the full map every frame, and a fast flyTo outruns tile loading
     // so almost nothing draws mid-flight. 6 s lets tiles keep up and stay smooth.
